Created seperate hood subsystem
